+++ core/app/c/[communitySlug]/pubs/[pubId]/components/jsonSchemaHelpers.tsx
@@ -8,7 +8,7 @@ import { cn } from "utils";
 
 import type { FileUpload } from "~/lib/fields/fileUpload";
 import type { PubField } from "~/lib/types";
-import { FileUploadPreview } from "~/app/components/Forms/FileUpload";
+import { FileUploadPreview } from "~/app/components/forms/FileUpload";
 
 interface PubFieldWithValue extends PubField {
 	schema: PubFieldSchema | null;
